+++ app/src/main/java/com/example/runtracker/database/RunViewModel.kt
@@ -7,7 +7,7 @@ class RunViewModel(private val repository: RunRepository): ViewModel() {
     var runs: LiveData<List<Run>> = repository.allRuns.asLiveData()
     var numberOfRuns: LiveData<Int> = repository.numberOfRuns.asLiveData()
     var totalDistance: LiveData<Float> = repository.totalDistance.asLiveData()
-    var totalDuration: LiveData<Long> = repository.totalDuration.asLiveData()
+    var totalDuration: LiveData<Int> = repository.totalDuration.asLiveData()
     //var totalPhotos: LiveData<Int> = repository.totalPhotos.asLiveData()
     var longestDistance: LiveData<Float> = repository.longestDistance.asLiveData()
     var longestDuration: LiveData<Int> = repository.longestDuration.asLiveData()
